[MSDK-7083] Move processing applicators to own package
