+++ standard-integration/app/src/main/java/io/mimi/example/android/applicators/processing/IsEnabledApplicator.kt
@@ -1,4 +1,4 @@
-package io.mimi.example.android.applicators
+package io.mimi.example.android.applicators.processing
 
 import io.mimi.example.android.MimiProcessingApplicator
 import io.mimi.sdk.core.controller.processing.MimiApplicatorResult
